Add npm prepublish support
With our current setup we were running into a couple issues when
importing juttle-viz into other libraries. 1) We couldn't import
and browserify the library because we weren't babelify the source
and 2) We weren't generating a javascript blob that I can throw
into my browser.

What I did here:
- Created `gulp lib` task that run all js code through babel and spits
  it out into a lib dir. As such I've pointed the entry to lib/index.js.
  Now requiring juttle-viz should "Just Work"(TM).
- From the above had to rename folder lib in src to ext so my .gitignore
  on lib wouldn't remove lib from src/lib. I don't think src should have a
  folder called lib anyways.
- Add a `gulp clean` command that removes generated dirs (lib, build and
  test/build)
- Add an npm "prepublish" script that cleans the repo then generates the
  babelified lib and browserify artifact.

Moving forward: Kind of weird that babel has two different libraries:
gulp-babel and gulp-babelify. Ideally we create the artifact and lib
using the same babel lib. In a future pr we might want to use webpack
while might give us the flexibility that we need.